fix price data type

+++ source/includes/fundamentals/code-examples/Clothing.cs
@@ -15,7 +15,7 @@ public class Clothing
     public bool InStock { get; set; }
 
     [BsonElement("price")]
-    [BsonRepresentation(BsonType.Double)]
+    [BsonRepresentation(BsonType.Decimal128)]
     public decimal Price { get; set; }
 
     [BsonElement("colorSelection")]
